WIP give up for now on monotonicity
Doing the proof this way is completely insane, Coq is too slow for it. What's
worse, the proof *almost* works by normal induction, except for the case for
recursive types: in that case, we must use complete induction using the
induction metric. I should check on paper if that would indeed work, and admit
the lemma for now.