Add bounds to type selection and drop later from its expansion
That's the most significant change to the model since quite a while.
Basic idea: while the actual definition of the abstract type is only accessible
in a later world, its bounds are usable in the current world, so we just need to
take advantage of that both in type selections and in the typing rule for type
members. This allows removing annoying uses of later from typing rules involving
type selections.

The only annoyance is that uses of later disappear also when establishing
subtyping of type member definitions, but since that hasn't been used much until
now, this shouldn't be too problematic.