dot_wfrel2.v: Build sem_vl_subtype into sem_subtype
Instead of trying to really prove `subtype_to_vl_subtype`, just make it hold by
construction by strengthening sem_subtype.

I did this just after realizing that vl_to_tm is very hard to prove, even with
let in values, because we can't ensure (for arbitrary values) that the closures
close over the same environments (which could be fixable).

Even ignoring that, the proof looks insanely hard because of the complexity of
the necessary inductive hypothesis --- it looks comparable to CPS/ANF conversion.